Remove :cheri from most conditional operations
Initially the plan was that the meta-codasip-cheri layer could be
present but disabled by not enabling the 'cheri' override (similar to
the way meta-codasip-baremetal behaves). However it now appears that
it would be useful to be able to build the same version and
configuration of packages with and without cheri to:
 - ensure that code modified for cheri still builds when it is disabled
 - for demonstration
 - for cross checking

So change the layer so that simply having it present changes the versions
and configurations to be those needed for cheri, but only enable cheri
compilation when the override is enabled.