musl: Delete header files which are also provided by the compiler
In general, when both the compiler and musl provide the same header
files, we want to use the header files provided by the compiler.
However because of the way Yocto specifies the target header files
using --sysroot, this doesn't currently work. So delete the musl
provided files from the install directory.