qemu: Remove x86_64 from qemu-system-native build
When x86_64 was added to QEMU_TARGETS in commit 3bae428a1 ("Add x86_64
to QEMU_TARGETS") this was needed for the dummy qemu builds, but also
accidental added it to the list of targets for qemu-system-native.
As don't need to support x86_64 in this case, remove it in the
qemu-system-native recipe.