busybox: apply two vi fixes
Apply two patches that fix crashes in busybox vi.

vi reads p[-1] where p points to the beginning of a line. This crashes if
we're on the first line.

The problem can be reproduced simply be starting vi and then
pressing O (or I) and ESC a couple of times (it might be necessary to
enter some text between O/I and ESC).

The patches should go into meta-codasip-cheri where we're using busybox
1.32. Newer busybox versions do already have fixes for this problem.