LTP: Improve dependencies, fix tests
Now that all dependencies (except perl and python) of the LTP
testsuite compile for CHERI use the upstream dependency list and
only remove perl and python from it.

Add some more dependencies that are needed to run tests.

Add fixes for some tests that fault with CHERI and remove a
previous workaround.

Add some quirks that make LTP more usable on qemu but make them
depend on the qemu-tweaks distro feature.

How to run the tests:
- Add "ltp" to your image.
- Start the image and run /opt/ltp/runltp to run the entire suite.
- Network tests are a separate suite and will need the "-N" option.