ltp: Fix clone302 test case and disable sbrk tests
The syscall clone3 passes an argument struct containing a (void *) to
the child's stack. In the ltp test, an (unsigned long) was used for the
stack address. However, as this is a capability, a (uintptr_t) shall be
used.

Furthermore, we disable the sbrk tests. As sbrk is a wrapper around brk
and brk is not supported on CHERI[1], these tests don't make sense.

[1] https://git.morello-project.org/morello/kernel/linux/-/wikis/Morello-pure-capability-kernel-user-Linux-ABI-specification#brk