acl: add patch for libacl to support 16-byte aligned pointer
I initially tried a container_of()-like approach to perform the
transformation from external pointer to internal pointer. This worked
fine for the ext2int macro where type information is available. But it
is not feasible for libacl's acl_int(void* obj_p) where we don't have
information about the enclosing type.

While the applied patch is not ideal, as it increases libacl's struct
sizes and therefore enlarges the libraries memory footprint, it is
probably the best we can do at the moment besides a larger refactoring
of libacl's code.

Fixes #1.