libdrm: Fix linux definition of drm_uptr_t
The code for the *BSD case correctly defines drm_uptr_t as a
capability sized object on 64-bit. For linux use the new __u64ptr
for this. While there fix another type that need to be a pointer
at least on linux by using "__uptr" instead of "unsigned long".
The latter will likely need some help for the BSD case.