strace: Apply (most) CHERI patches for non-CHERI builds
To test the strace changes for CHERI still build for non-CHERI
systems, apply the patches in all cases. However there are still some
changes which are CHERI specific, so the UAPI patch is split into a
part that is CHERI specific and a part that is applied in all cases.